+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [0.0.29] 2021-04-14
 ### Changed
-- Integrate with wutsi-stream-rabbitmq 1.3.0 to support DLQ replay
+- Integrate with wutsi-stream-rabbitmq 1.4.0 to support DLQ replay
 ### Fixed
 - API interface with query string